Change project Sdk moniker  to Microsoft.NET.Sdk for all packable projects Fix bugs with Swagger and Assets path
