+++ src/VirtoCommerce.Platform.Assets.AzureBlobStorage/VirtoCommerce.Platform.Assets.AzureBlobStorage.csproj
@@ -4,7 +4,7 @@
         <TargetFramework>netcoreapp3.1</TargetFramework>
         <noWarn>1591</noWarn>
         <OutputType>Library</OutputType>
-		<IsPackable>False</IsPackable>
+		<IsPackable>True</IsPackable>
     </PropertyGroup>
 
     <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
