+++ src/VirtoCommerce.Platform.Data/Extensions/ServiceCollectionExtensions.cs
@@ -1,15 +1,12 @@
 using System;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
-using VirtoCommerce.Platform.Core;
+using VirtoCommerce.Platform.Caching;
 using VirtoCommerce.Platform.Core.Bus;
-using VirtoCommerce.Platform.Core.Caching;
 using VirtoCommerce.Platform.Core.ChangeLog;
-using VirtoCommerce.Platform.Core.Common;
 using VirtoCommerce.Platform.Core.Events;
 using VirtoCommerce.Platform.Core.ExportImport;
 using VirtoCommerce.Platform.Core.Localizations;
@@ -20,10 +17,8 @@ using VirtoCommerce.Platform.Data.ChangeLog;
 using VirtoCommerce.Platform.Data.DynamicProperties;
 using VirtoCommerce.Platform.Data.ExportImport;
 using VirtoCommerce.Platform.Data.Localizations;
-using VirtoCommerce.Platform.Data.PushNotifications;
 using VirtoCommerce.Platform.Data.Repositories;
 using VirtoCommerce.Platform.Data.Settings;
-using VirtoCommerce.Platform.Caching;
 
 namespace VirtoCommerce.Platform.Data.Extensions
 {
@@ -40,8 +35,6 @@ namespace VirtoCommerce.Platform.Data.Extensions
 
             services.AddDynamicProperties();
 
-            services.AddSingleton<IPushNotificationManager, PushNotificationManager>();
-
             var inProcessBus = new InProcessBus();
             services.AddSingleton<IHandlerRegistrar>(inProcessBus);
             services.AddSingleton<IEventPublisher>(inProcessBus);
@@ -55,14 +48,7 @@ namespace VirtoCommerce.Platform.Data.Extensions
 
             services.AddTransient<IEmailSender, DefaultEmailSender>();
 
-             // TODO: (AK) 
-            services.AddSingleton(js =>
-            {
-                var serv = js.GetService<IOptions<MvcNewtonsoftJsonOptions>>();
-                return JsonSerializer.Create(serv.Value.SerializerSettings);
-            });
-            
-
+         
             //Register dependencies for translation
             services.AddSingleton<ITranslationDataProvider, PlatformTranslationDataProvider>();
             services.AddSingleton<ITranslationDataProvider, ModulesTranslationDataProvider>();
