--- src/VirtoCommerce.Platform.Modules/Extensions/ApplicationBuilderExtensions.cs
@@ -1,39 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-using VirtoCommerce.Platform.Core.Modularity;
-
-namespace VirtoCommerce.Platform.Modules.Extensions
-{
-    public static class ApplicationBuilderExtensions
-    {
-        public static IApplicationBuilder UseModules(this IApplicationBuilder appBuilder)
-        {
-            using (var serviceScope = appBuilder.ApplicationServices.CreateScope())
-            {
-                var moduleManager = serviceScope.ServiceProvider.GetRequiredService<IModuleManager>();
-                var modules = GetInstalledModules(serviceScope.ServiceProvider);
-                foreach (var module in modules)
-                {
-                    moduleManager.PostInitializeModule(module, appBuilder);
-                }
-            }
-            return appBuilder;
-        }
-
-        private static IEnumerable<ManifestModuleInfo> GetInstalledModules(IServiceProvider serviceProvider)
-        {
-            var moduleCatalog = serviceProvider.GetRequiredService<ILocalModuleCatalog>();
-            var allModules = moduleCatalog.Modules.OfType<ManifestModuleInfo>()
-                                          .Where(x => x.State == ModuleState.Initialized && !x.Errors.Any())
-                                          .ToArray();
-
-            return moduleCatalog.CompleteListWithDependencies(allModules)
-                .OfType<ManifestModuleInfo>()
-                .Where(x => x.State == ModuleState.Initialized && !x.Errors.Any())
-                .ToArray();
-        }
-    }
-}
