+++ src/VirtoCommerce.Platform.Web/Security/CustomUserManager.cs
@@ -13,7 +13,7 @@ using VirtoCommerce.Platform.Core.Security;
 using VirtoCommerce.Platform.Core.Security.Events;
 using VirtoCommerce.Platform.Security.Caching;
 
-namespace VirtoCommerce.Platform.Security.Services
+namespace VirtoCommerce.Platform.Web.Security
 {
     public class CustomUserManager : AspNetUserManager<ApplicationUser>
     {
