+++ src/VirtoCommerce.Platform.Web/VirtoCommerce.Platform.Web.csproj
@@ -38,6 +38,7 @@
         <PackageReference Include="OpenIddict.EntityFrameworkCore" Version="2.0.1" />
         <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc5" />
         <PackageReference Include="Swashbuckle.AspNetCore.Cli" Version="4.0.1" />
+        <PackageReference Include="Swashbuckle.AspNetCore.Filters" Version="5.0.0-rc9" />
         <PackageReference Include="System.Security.Cryptography.Algorithms" Version="4.3.1" />
     </ItemGroup>
     <ItemGroup>
